Handle time type
