+++ packages/strapi-plugin-content-type-builder/controllers/validation/content-type.js
@@ -33,6 +33,8 @@ const VALID_TYPES = [
   'float',
   'decimal',
   'date',
+  'time',
+  'datetime',
   'boolean',
 
   // nested component
